<0001f9ff> Add gas snapshot, npm linting, remappings, and snippets

+++ src/BasicVault.sol
@@ -1,34 +1,34 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.13;
 
-import 'solmate/tokens/ERC20.sol';
+import "solmate/tokens/ERC20.sol";
 
 contract BasicVault {
-    
+    //
     ERC20 public immutable token;
-    mapping(address => uint) public balances;
+    mapping(address => uint256) public balances;
 
-    event Deposit(address indexed from, uint amount);
-    event Withdrawal(address indexed from, uint amount);
+    event Deposit(address indexed from, uint256 amount);
+    event Withdrawal(address indexed from, uint256 amount);
 
     constructor(ERC20 token_) {
         token = token_;
     }
 
-    function deposit(uint amount) external {
+    function deposit(uint256 amount) external {
         balances[msg.sender] += amount;
 
         bool success = token.transferFrom(msg.sender, address(this), amount);
-        require(success, 'Deposit failed!');
+        require(success, "Deposit failed!");
 
         emit Deposit(msg.sender, amount);
     }
 
-    function withdraw(uint amount) external {
+    function withdraw(uint256 amount) external {
         balances[msg.sender] -= amount;
 
         bool success = token.transfer(msg.sender, amount);
-        require(success, 'Withdrawal failed!');
+        require(success, "Withdrawal failed!");
 
         emit Withdrawal(msg.sender, amount);
     }
